main: refactor `OnGetData` to simplify the logic
The preivous usage of two channels - by piping them together to create a
semaphore effect, is difficult to follow and prune to bugs. This commit
now refactors the method to explicitly implement a semaphore.

Prior to this change, we would allow at max 3 concurrent goroutines -
this is now bumped to 5.