txscript: create EC op code test vector gen infra
In this commit, we create some test vector infra for the EC op tests.
With this, we'll use a build tag to control if we write certian unit
tests as test vectors on disk or not. If we're generating test vectors,
then we make sure to always generate the same set of keys each time.