blockchain: change HaveBlock to also check for block data availability
In block processing and block downloading, HaveBlock is used to check if
the block data already exists.  It was ok to just check for the
existence of the blockNode but we now we also need to check if the data
exists as the blockNode may be present for just the block header.