mempool: add MaxStandardTxVersion constant for v3 transaction support
This commit introduces the MaxStandardTxVersion constant to the mempool
policy package, setting it to 3 to enable support for version 3
transactions. Previously, the maximum accepted transaction version was
hardcoded to 2 in the server configuration.

Note that we don't yet implement the new TRUC rules, yet. But this'll
enable acceptance into the btcd mempool.