silentpayments: add functions for block tweaks
This commit adds library functions for calculating block tweaks. A
block tweak is the set of Silent Payment transaction tweaks that are
contained in a block. A SP transaction tweak is the pre-calculated input
public key sum tweaked by the input hash of a transaction and is
required for light clients to be able to calculate potential Silent
Payment outputs in a transaction or block without needing to download
the full block/transaction.